UEye_Camera: started the implementation of the 'stop' method.
Changed the prototype of the 'start' method to get the ring buffer size
as a size_t type variable.

+++ include/ueye_camera.hpp
@@ -130,7 +130,7 @@ class UEye_Camera: public Camera {
         /**
          * @brief Starts image acquisition. 
          */
-        virtual void start(Image* ringBuffer[], unsigned int bufferSize); 
+        virtual void start(Image* ringBuffer[], size_t bufferSize); 
 
         /**
          * @brief Stops the image acquisition.
